+++ hugegraph-client/src/main/java/com/baidu/hugegraph/structure/traverser/PathsRequest.java
@@ -39,7 +39,7 @@ public class PathsRequest {
     @JsonProperty("capacity")
     public long capacity = Traverser.DEFAULT_CAPACITY;
     @JsonProperty("limit")
-    public long limit = Traverser.DEFAULT_LIMIT;
+    public int limit = Traverser.DEFAULT_LIMIT;
     @JsonProperty("with_vertex")
     public boolean withVertex = false;
 
@@ -102,7 +102,7 @@ public class PathsRequest {
             return this;
         }
 
-        public PathsRequest.Builder limit(long limit) {
+        public PathsRequest.Builder limit(int limit) {
             TraversersAPI.checkLimit(limit);
             this.request.limit = limit;
             return this;
